작업사항(21/10/24)
1. Component
Script -> CreateGameObject, Destroy, Instantiate 함수

+++ Project/DX11_2D_GameEngine_Lib/EventManager.cpp
@@ -38,7 +38,7 @@ void EventManager::Excute(const EventStruct& event_struct)
 		SceneManager::GetInstance()->GetCurrentScene()->AddGameObject
 		(
 			p_new_game_object,
-			static_cast<UINT>(event_struct.layer_index),
+			event_struct.layer_index,
 			false
 		);
 	}
